commit 3

+++ src/main/java/blog/app/Blog/repositories/PostRepository.java
@@ -4,8 +4,11 @@ import blog.app.Blog.models.entities.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface PostRepository extends JpaRepository<Post, Integer> {
+public interface PostRepository extends JpaRepository<Post, Long> {
+  //  Post deleteById (Long id);
